changed initial guess method to no longer use photutils. It now uses the center of the aperture as the intial guess for the center of the object
Added a routine for calculating the normalized chi squared value, for estimating the goodness of fit

Finished the routine for analysing objects selected in DS9. The routine uses the selected regions as apertures, and attemps to fit a bivariant normal distrobution (2d gaussian).  It then plots the aperature, the mask used to shield the object, and if it found a fit, the width and center of the gaussian.