updated code to use pyds9.get_fits() from pyds9.get_pyfits()
Found a serious bug in get_regions(). Sometime, the region definition does not include a dot '.', which was not accounted for in the re pattern. Updated get_region to actually find the total number, with an optional decimal.

Also changed the region definitions in class Region() to be floats from ints. This is to prevent propagation of rounding errors