Did a bunch of testing with the sum of two Moffats model on a fake object.
I was not able to reproduce the problems that had been apparant before,
so I created a test script for fitting a real object. It manged to
generate a fit, but it is way off.

starting using file k4m_160531_050920_ori.fits.fz as my real object
source, because it has much less eccentricity.

Tried to improve numerical normalization, but it broke